feat: make svg folder
