+++ svgs/CloseButton.tsx
@@ -1,0 +1,21 @@
+
+export const CloseButton = () => {
+    return (
+      <svg
+        width="20"
+        height="21"
+        viewBox="0 0 20 21"
+        fill="none"
+        xmlns="http://www.w3.org/2000/svg"
+      >
+        <path
+          d="M15.8337 4.6665L4.16699 16.3332M4.16699 4.6665L15.8337 16.3332"
+          stroke="#F9F9F9"
+          strokeWidth="1.5"
+          strokeLinecap="round"
+          strokeLinejoin="round"
+        />
+      </svg>
+    );
+  };
+
